Feat: 갤러리리스트페이지 수정
